% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMSP.R
\docType{data}
\name{AMSP}
\alias{AMSP}
\title{National River Flow Archive (NRFA) annual maximum data for sites suitable for pooling}
\format{
A data frame with 25669 rows and 3 columns
\describe{
  \item{Date}{Date}
  \item{Flow}{Annual maximum peak flow, m3/s}
  \item{id}{Station identification number}
}
}
\source{
\url{https://nrfa.ceh.ac.uk/peak-flow-dataset}
}
\usage{
AMSP
}
\description{
A data.frame with three columns; Date, Flow, id. NRFA Peak Flow Dataset - Version 11.1
}
\keyword{datasets}
