% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{GetAM}
\alias{GetAM}
\title{Get an annual maximum sample from the National River Flow Archive sites suitable for pooling}
\usage{
GetAM(ref)
}
\arguments{
\item{ref}{the site reference of interest (numeric)}
}
\value{
A data.frame with columns; Date, Flow, and id
}
\description{
Extracts the annual maximum peak flow sample and associated dates for the site of interest.
}
\examples{
#Get an AMAX sample and display it in the console
GetAM(203018)
#Save an AMAX sample as an object
AM.203018 <- GetAM(203018)
}
\author{
Anthony Hammond
}
