% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{GetQMED}
\alias{GetQMED}
\title{QMED from a gauged site suitable for QMED}
\usage{
GetQMED(x)
}
\arguments{
\item{x}{the gauged reference}
}
\value{
the median annual maximum
}
\description{
Provides QMED (median annual maximum flow) from a site suitable for QMED, using the site reference.
}
\examples{
#Get the observed QMED from sites 55002
GetQMED(55002)
}
\author{
Anthony Hammond
}
