% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{Kappa3Est}
\alias{Kappa3Est}
\title{Kappa3 distribution estimates from parameters}
\usage{
Kappa3Est(loc, scale, shape, q = NULL, RP = 100)
}
\arguments{
\item{loc}{location parameter}

\item{scale}{scale parameter}

\item{shape}{shape parameter}

\item{q}{quantile. magnitude of the variable under consideration}

\item{RP}{return period}
}
\value{
quantile as a function of RP or vice versa
}
\description{
Estimated quantiles as function of return period (RP) and vice versa, from user input parameters
}
\details{
If the argument q is used, it overrides RP and provides RP as a function of q (magnitude of variable) as opposed to q as a function of RP. This is the Kappa3 distribution as defined in Kjeldsen, T (2019), 'The 3-parameter Kappa distribution as an alternative for use with FEH pooling groups.'Circulation - The Newsletter of the British Hydrological Society, no. 142.
}
\examples{
#Get an annual maximum sample, estimate the parameters and estimate 50-year RP
AM.27090 <- GetAM(27090)
#Get parameters and Store as an object
Pars <- as.numeric(Kappa3Pars(AM.27090$Flow))
#get estimate of 50-yr flow
Kappa3Est(Pars[1], Pars[2], Pars[3], RP = 50)
#Estimate the RP for a 600m3/s discharge
Kappa3Est(Pars[1], Pars[2], Pars[3], q = 600)
}
\author{
Anthony Hammond
}
