% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{Lmoms}
\alias{Lmoms}
\title{Lmoments & Lmoment ratios}
\usage{
Lmoms(x)
}
\arguments{
\item{x}{a numeric vector. The sample of interest}
}
\value{
A data.frame with one row and column headings; L1, L2, L3, L4, Lcv, LSkew, and LKurt. The first four are the Lmoments and the next three are the Lmoment ratios.
}
\description{
Calculates the Lmoments and Lmoment ratios from a sample of data
}
\details{
Lmoments calculated according to methods outlined by Hosking & Wallis (1997): Regional Frequency Analysis and approach based on LMoments. Also in the Flood Estimation Handbook (1999), volume 3.
}
\examples{
#Get an AMAX sample and calculate the Lmoments
AM.96001 <- GetAM(96001)
Lmoms(AM.96001$Flow)
}
\author{
Anthony Hammond
}
