% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEH08.R
\name{Pool_FEH08}
\alias{Pool_FEH08}
\title{Create pooling group using the FEH08 method}
\usage{
Pool_FEH08(
  CDs,
  N = 500,
  UrbMax = 0.03,
  DeUrb = TRUE,
  exclude = NULL,
  include = NULL
)
}
\arguments{
\item{CDs}{catchment descriptors derived from either GetCDs or CDsXML}

\item{N}{minimum Number of total gauged record years for the pooling group}

\item{UrbMax}{Maximum URBEXT2000 level with a default of 0.03. Any catchment with URBEXT2000 above this level will be excluded from the pooling group}

\item{DeUrb}{logical argument with a default of TRUE. If TRUE, the LCVs of all sites in the pooling group are "De-Urbanised" according to the FEH08 method.}

\item{exclude}{sites to exclude from the pooling group. Either a single site reference or a vector of site references (numeric). If this is used the next site with the lowest SDM is included such that the total sample of AMAX is at least N.}

\item{include}{sites to include that otherwise would not be included by default. For example if it is a subject site that has URBEXT2015 above UrbMax. Or one that has not been selected automatically using the similarity distance measure.}
}
\value{
A data.frame of the pooling group with site reference row names and 14 columns, each providing catchment & gauge details for the sites in the pooling group.
}
\description{
Legacy function to develop a pooling group based on catchment descriptors
}
\details{
A pooling group is created from a CDs object, derived from GetCDs or CDsXML (CDsXML_Legacy). To change the default pooling group, one or more sites can be excluded using the 'exclude' option, which requires either a site reference or multiple site references in a vector. If this is done, the site with the next lowest similarity distance measure is added to the group (until the total number of years is at least N). Sites with URBEXT2000 (urban extent) > 0.03 are excluded from the pooling group by default. This threshold can be adjusted with UrbMax. If DeUrb is set as TRUE (the default), the LCV and LSKEW values for sites in the pooling group are de-urbanised.

The pooling method is as specified by FEH2008. Accordingly the SAAR applied is SAAR6190, FARL is FARL (as opposed to FARL2015) and BFIHOST is BFIHOST (as opposed to BFIHOST19 or BFIHOST19scaled)
}
\examples{
# Get some catchment descriptors
cds_73005 <- GetCDs(73005)

# Set up a pooling group object called pool_73005 excluding sites 79005 & 46003
# Then print the group to the console
pool_73005 <- Pool_FEH08(cds_73005, exclude = c(79005, 46003))
pool_73005


}
\author{
Anthony Hammond
}
