% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{PoolSmall}
\alias{PoolSmall}
\title{Create pooling group for small catchments}
\usage{
PoolSmall(
  CDs = NULL,
  AREA,
  SAAR,
  N = 500,
  exclude = NULL,
  iug = FALSE,
  DeUrb = FALSE
)
}
\arguments{
\item{CDs}{catchment descriptors derived from either GetCDs or CDsXML}

\item{AREA}{catchment area in km2}

\item{SAAR}{catchment standard average annual rainfall (1961-1990) in mm}

\item{N}{minimum Number of total gauged record years for the pooling group}

\item{exclude}{sites to exclude from the pooling group. Either a single site reference or a vector of site references (numeric)}

\item{iug}{iug stands for 'include urban gauge'. It's a logical argument with default of FALSE. TRUE will over-ride the default and add the closest site in catchment descriptor space to the CDs provided if it has URBEXT2000 >= 0.03}

\item{DeUrb}{logical argument with a default of FALSE. If true, the Lcv and LSkew of the top site in the pooling group will be de-urbanised}
}
\value{
A data.frame of the pooling group with site reference row names and 24 columns, each providing catchment & gauge details for the sites in the pooling group.
}
\description{
Function to develop a small catchments pooling group based on catchment descriptors
}
\details{
A pooling group is created from a CDs object, derived from GetCDs or CDsXML, or specifically with the necessary catchment descriptors (see arguments). To change the default pooling group one or more sites can be excluded using the 'exclude' option, which requires either a site reference or multiple site references in a vector. If this is done, the site with the next lowest similarity distance measure is added to the group (until the total number of years is at least N). Sites with URBEXT2000 (urban extent) > 0.03 are excluded by default. If a gauged assessment is required and the site of interest is urban it can be included by setting iug = TRUE. In which case the user may wish to de-urbanise the subject site's Lcv and Lskew (L-moment ratios) by setting DeUrb = TRUE. If the user has more data available for a partcicular site within the pooling group, the Lcv and Lskew for the site can be updated after the group has been finalised.
}
\examples{
#Get some catchment descriptors
CDs.21001 <- GetCDs(21001)
#Set up a pooling group object called Pool.21001 excluding site 206006
#Then print the group to the console
Pool.21001 <- PoolSmall(CDs.21001, exclude = 206006)
Pool.21001
#Form a pooling group, called PoolGroup, with the catchment descriptors specifically
PoolGroup <- PoolSmall(AREA = 22, SAAR = 1702)
}
\author{
Anthony Hammond
}
