% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainAPI.R
\name{RainEA}
\alias{RainEA}
\title{Get Environment Agency rainfall data (England).}
\usage{
RainEA(
  Lat = 54,
  Lon = -2,
  Range = 10,
  WISKI_ID = NULL,
  Period = 0.25,
  DateSt = "2015-12-01",
  DateEnd = "2015-12-31"
)
}
\arguments{
\item{Lat}{Latitude of the point of interest. Provided when the user wants information about available local rain gauges}

\item{Lon}{Longitude of the point of interest. Provided when the user wants information about available local rain gauges}

\item{Range}{The radius (km) from the point of interest (Lat, Lon) for which the user wants rain gauge information.}

\item{WISKI_ID}{The WISKI identification (as "character") for the rain gauge of interest}

\item{Period}{The sampling rate of the rainfall in hours (generally 0.25 or 24. Default is 0.25 (15 minutes).}

\item{DateSt}{The start date of the data extraction in the form of "YYYY-MM-DD".}

\item{DateEnd}{The end date of the data extraction in the form of "YYYY-MM-DD".}
}
\value{
A data.frame with POSIXct (DateTime) in the first column, and rainfall (mm) in the second column. Unless the WISKI_ID is null, then information about rain gauges approximately within a radius (Range) of the Lat and Lon is povided.
}
\description{
Extract rainfall data from the Environment Agency's API.
}
\details{
The function provides one of two outputs. Either information about available local rain gauges, or the data from a specified gauge (specified by WISKI ID). The process is to find the local information (including WISKI ID) by using the latitude and longitude and range. Then use the WISKI ID to get the data. If data requested is not available, for example - outside the date range or not available at the requested sampling rate, an error message is returned stating "no lines available in input". There is currently a limit to how many lines of data can be extracted (100,000) for each use. That is approximately 274 years of daily rainfall, but only 2.85 years of 15 minute data.
}
\examples{
#Get information about available rain gauges.
#within a 10km radius of Lat = 54.5, Lon = -3.2
\dontrun{ RainEA(Lat = 54.5, Lon = -3.2) }
#Now we'll use the WISKI reference for the Honister rain gauge
# to get some data for the Dec 2015 (default dates)
\dontrun{ Honister15 <- RainEA(WISKI_ID = "592463") }
#Now we'll have a look at the top of the data and plot it
\dontrun{ head(Honister15) }
\dontrun{ plot(Honister15, type = "h", ylab = "Rainfall (mm)") }
#Now we'll get the daily rain for the same period and plot it
\dontrun{ Honister24 <- RainEA(WISKI_ID = "592463", Period = 24) }
\dontrun{ plot(Honister24, type = "h", ylab = "Rainfall (mm)") }
}
\author{
Anthony Hammond
}
