% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThamesPQ.R
\docType{data}
\name{ThamesPQ}
\alias{ThamesPQ}
\title{Kingston upon Thames daily flow and catchment precipitation 2000-10-01 to 2015-09-30}
\format{
A data frame with 5478 rows and 3 columns:
\describe{
  \item{Date}{Date}
  \item{P}{Precipitation, in mm}
  \item{Q}{Daily mean discharge, in m3/s}
}
}
\source{
\url{https://nrfa.ceh.ac.uk/data/station/meanflow/39001}
}
\usage{
ThamesPQ
}
\description{
A data.frame of three columns; Date, Precipitation (P), & daily mean flow (Q)
}
\keyword{datasets}
