% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKOutline.R
\docType{data}
\name{UKOutline}
\alias{UKOutline}
\title{UK outline}
\format{
A data frame with 3867 rows and 2 variables
\describe{
  \item{X_BNG}{Easting, British national grid reference}
  \item{Y_BNG}{Northing, British national grid reference}
}
}
\source{
\url{https://environment.data.gov.uk/}
}
\usage{
UKOutline
}
\description{
Easting and northing national grid reference points around the coast of the UK
}
\keyword{datasets}
