% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{WeightsUnLcv}
\alias{WeightsUnLcv}
\title{Site ungauged linear coefficient of variation (Lcv) weightings}
\usage{
WeightsUnLcv(x)
}
\arguments{
\item{x}{pooling group derived with the Pool() function}
}
\value{
A data.frame with site references in the first column and associated weights in the second
}
\description{
Provides the ungauged Lcv weights for each site in a pooling group
}
\details{
Weighting method as according to Science Report: SC050050 - Improving the FEH statistical procedures for flood frequency estimation
}
\examples{
# Get some CDs, form an ungauged pooling group, and estimate ungauged Lcv
cds_27051 <- GetCDs(27051)
pool_27051 <- Pool(cds_27051, exclude = 27051)
WeightsUnLcv(pool_27051)

}
\author{
Anthony Hammond
}
