% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMPF.R
\docType{data}
\name{AMPF}
\alias{AMPF}
\title{National River Flow Archive (NRFA) annual maximum peak flow data for sites suitable for QMED only, and those for pooling.}
\format{
A data frame with 44474 rows and 3 columns
\describe{
  \item{Date}{Date}
  \item{Flow}{Annual maximum peak flow, m3/s}
  \item{id}{Station identification number}
}
}
\source{
\url{https://nrfa.ceh.ac.uk/data/peak-flow-dataset}
}
\usage{
AMPF
}
\description{
A data.frame with three columns; Date, Flow, id. NRFA Peak Flow Dataset - Version (15)
}
\keyword{datasets}
