% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{AMplot}
\alias{AMplot}
\title{Plot of the annual maximum sample}
\usage{
AMplot(x)
}
\arguments{
\item{x}{a data.frame with at least two columns. The first a date column and the second the annual maximum (AM) sequence. A third column with the station id is necessary for inclusion of the id in the plot title. An AM object derived from the GetAM or AMImport functions.}
}
\value{
A barplot of the annual maximum sample
}
\description{
Provides two plots. First, a histogram of the sample, second, a barplot
}
\details{
When used with a GetAM object or any data.frame with dates in the first column, the barplot is daily. Therefore, although it's an annual maximum (AM) sequence, some bars will be closer together depending on the number of days between them.
}
\examples{
#Get an AMAX sample and plot
AMplot(GetAM(58002))
}
\author{
Anthony Hammond
}
