% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{CDsXML}
\alias{CDsXML}
\title{Import catchment descriptors from .xml files}
\usage{
CDsXML(x)
}
\arguments{
\item{x}{the xml file path}
}
\value{
A data.frame with columns; Descriptor and Value.
}
\description{
Imports catchment descriptors from xml files either from an FEH webservice download or from the Peakflows dataset downloaded from the national river flow archive (NRFA) website
}
\examples{

# Import catchment descriptors from a FEH webserver XML file and display XML in the console
\dontrun{
cds_my_site <- CDsXML(r"{C:\Data\FEH_Catchment_384200_458200.xml}")
cds_my_site
}

}
\author{
Anthony Hammond
}
