% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{GetCDs}
\alias{GetCDs}
\title{Get catchment descriptors from the National River Flow Archive sites considered suitable for median annual maximum flow estimation (QMED) and pooling.}
\usage{
GetCDs(x)
}
\arguments{
\item{x}{the site reference of interest (numeric)}
}
\value{
A data.frame with columns; Descriptor and Value.
}
\description{
Extracts the catchment descriptors for a site of interest from those suitable for QMED and pooling.
}
\examples{
#Get CDs and display in the console
CDs.203018 <- GetCDs(203018)
CDs.203018
}
\author{
Anthony Hammond
}
