% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetDataSEPA_Rain}
\alias{GetDataSEPA_Rain}
\title{Get Scottish Environment Protection Agency (SEPA) hourly rainfall data.}
\usage{
GetDataSEPA_Rain(
  Lat = NULL,
  Lon = NULL,
  Range = 30,
  StationName,
  From = NULL,
  To = NULL
)
}
\arguments{
\item{Lat}{Latitude of the point of interest. Provided when the user wants information about available local rain gauges}

\item{Lon}{Longitude of the point of interest. Provided when the user wants information about available local rain gauges}

\item{Range}{The radius (km) from the point of interest (Lat, Lon) for which the user wants a list of rain gauges (default is 30).}

\item{StationName}{The name of the station for which you want rainfall. If you type something other than one of the available stations, the list of stations will be returned.}

\item{From}{A start date for the data in the form of "YYYY-MM-DD". Default of NULL means the earliest available date is used}

\item{To}{An end date for the data in the form of "YYYY-MM-DD". The default is the most recent date available.}
}
\value{
A data.frame with POSIXct in the first column, and rainfall in the second column. Unless the StationName provided is not in the available list, then the available list is returned.
}
\description{
Extract hourly rainfall data from SEPA's API.
}
\details{
You can download data using the gauge name and you can find gauges within a given range using the latitude and longitude. If the 'From' date is left as null, the earliest date of available data will be used. If the 'To' date is left as null, the most recent date of available data will be used.
}
\examples{
# Get the list of available stations
\dontrun{
GetDataSEPA_Rain(StationName = "AnythingButAStationName")
}

# Get rain from the Bannockburn station
\dontrun{
bannockburn <- GetDataSEPA_Rain(
  StationName = "Bannockburn",
  From = "1998-10-01", To = "1998-10-31"
)
}

# Inspect the first few rows and plot the data
\dontrun{
head(bannockburn)
plot(bannockburn, type = "h", ylab = "Rainfall (mm)")
}

}
\author{
Anthony Hammond
}
