% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFlowData.R
\docType{data}
\name{PeakFlowData}
\alias{PeakFlowData}
\title{National River Flow Archive descriptors and calculated statistics for sites suitable for QMED & pooling}
\format{
A data frame with 902 rows and 36 variables
}
\source{
\url{https://nrfa.ceh.ac.uk/data/peak-flow-dataset}
}
\usage{
PeakFlowData
}
\description{
A data.frame of catchment & data descriptors. NRFA Peak Flow Dataset - Version 14
}
\details{
Most of the FEH statistical method functions rely on this data. However, the data frame is open for manipulation in case the user wishes to add sites that aren't included, or change parts where local knowledge has improved on the data. If changes are made, they will only remain within the workspace. If a new workspace is opened and the UKFE package is loaded, the data frame will have returned to it's original state.
}
\keyword{datasets}
