% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All.R
\name{Zdists}
\alias{Zdists}
\title{Zdist Goodness of fit measure for pooling groups}
\usage{
Zdists(x)
}
\arguments{
\item{x}{pooling group derived from the Pool() function}
}
\value{
A list with the first element a data.frame of two GoF scores related to the columns; "GEV and "GenLog". The second element is a character stating which has the best fit.
}
\description{
Calculates the goodness of fit score for pooling groups with the method outlined in the Flood Estimation Hanbook (1999), volume 3.
}
\details{
The goodness of fit measure was developed by Hosking & Wallis and can be found in their book 'Regional Frequency Analysis: an approach based on LMoments (1997), as well as Flood Estimation Handbook volume 3.
}
\examples{
#Get CDs, form a pooling group and calculate the Zdist
CDs.203018 <- GetCDs(203018)
Pool.203018 <- Pool(CDs.203018)
Zdists(Pool.203018)
}
\author{
Anthony Hammond
}
